# src/openapi.yaml
openapi: '3.1.0'
info:
  title: WIZBI Control Plane API
  version: '1.0.0'
  description: |
    REST API for the WIZBI Control Plane — manages organizations, projects, 
    templates, users, and GCP/GitHub infrastructure provisioning.
    
    **Authentication**: All endpoints require either:
    - `X-Firebase-ID-Token` header (for web UI)
    - `X-API-Key` header (for programmatic/agent access)
    
    **MCP Server**: Available at `/api/mcp/sse` for AI agent integration.
  license:
    name: MIT
servers:
  - url: /api
    description: Local / deployed server
components:
  securitySchemes:
    firebaseToken:
      type: apiKey
      in: header
      name: X-Firebase-ID-Token
      description: Firebase ID Token for web UI auth
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for programmatic access
...
security:
  - firebaseToken: []
  - apiKey: []
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      responses:
        '200':
          description: List of projects
    post:
      tags: [Projects]
      summary: Create project and trigger provisioning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orgId, displayName, shortName, template]
              properties:
                orgId: { type: string }
                displayName: { type: string }
                shortName: { type: string }
                template: { type: string }
      responses:
        '201':
          description: Project created, provisioning started
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Get project details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Project details
    delete:
      tags: [Projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Deletion started
  /projects/{id}/logs:
    get:
      tags: [Projects]
      summary: Get provisioning logs
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: List of log entries
  /projects/{id}/provision:
    post:
      tags: [Projects]
      summary: Retry provisioning
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Provisioning restarted
  /projects/{id}/links:
    post:
      tags: [Projects]
      summary: Add external link to project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, name, color, icon]
              properties:
                url: { type: string }
                name: { type: string }
                color: { type: string }
                icon: { type: string }
      responses:
        '201':
          description: Link added
  /projects/{projectId}/links/{linkId}:
    delete:
      tags: [Projects]
      summary: Remove link from project
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: linkId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Link removed

  # ─── Global Links ──────────────────────────────────
  /global-links:
    get:
      tags: [Links]
      summary: Get global links
      responses:
        '200':
          description: List of global links
    post:
      tags: [Links]
      summary: Add global link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, name, color, icon]
              properties:
                url: { type: string }
                name: { type: string }
                color: { type: string }
                icon: { type: string }
      responses:
        '201':
          description: Global link added
  /global-links/{linkId}:
    delete:
      tags: [Links]
      summary: Remove global link
      parameters:
        - name: linkId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Global link removed

  # ─── Templates ─────────────────────────────────────
...